# docs/swagger.yaml
definitions:
  entities.Model3d:
    properties:
      create_date:
        type: string
      description:
        type: string
      file_path:
        type: string
      fileId:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
  server.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  server.HTTPResponse:
    properties:
      code:
        example: 200
        type: integer
      message:
        type: any
    type: object
info:
  contact: {}
paths:
  /models:
    get:
      description: Get list models
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Model3d'
        "204":
          description: Empty
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Get models
      tags:
      - Models
  /models/{id}:
    delete:
      description: Delete model
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.HTTPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Delete model
      tags:
      - Models
    get:
      description: Get model by ID
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Model3d'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Get model
      tags:
      - Models
    put:
      description: Edit model
      parameters:
      - description: Model Name
        in: formData
        name: name
        required: true
        type: string
      - description: Model Description
        in: formData
        name: descr
        type: string
      - description: Model File
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.HTTPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Edit model
      tags:
      - Models
  /models/create:
    post:
      description: Create model
      parameters:
      - description: Model Name
        in: formData
        name: name
        required: true
        type: string
      - description: Model Description
        in: formData
        name: descr
        type: string
      - description: Model File
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.HTTPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Create model
      tags:
      - Models
swagger: "2.0"